requirements.txt
# Framework agentique principal
agno[os]>=1.0.0

# Modèle LLM — Groq (gratuit)
# Clé API gratuite sur : https://console.groq.com
groq>=0.9.0

# Serveur web (pour exposer l'AgentOS)
fastapi>=0.110.0
uvicorn>=0.29.0

# Gestion des variables d'environnement
python-dotenv>=1.0.0